requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.16.1
plotly_express==0.4.1
pillow==9.4.0
python-barcode==0.15.1
google-cloud==0.34.0
google-cloud-storage==2.13.0
sendgrid==6.10.0
python-dotenv==1.0.0
pyDrive==1.3.1
